+++ renku/command/view_model/plan.py
@@ -20,15 +20,14 @@
 from __future__ import annotations
 
 import json
-from typing import TYPE_CHECKING, List, Optional, Union
+from typing import List, Optional, Union
 
-from renku.domain_model.provenance.agent import Person
+from renku.command.view_model.agent import PersonViewModel
 from renku.domain_model.workflow.composite_plan import CompositePlan
 from renku.domain_model.workflow.parameter import CommandInput, CommandOutput, CommandParameter
 from renku.domain_model.workflow.plan import AbstractPlan, Plan
 
-if TYPE_CHECKING:
-    from .composite_plan import CompositePlanViewModel
+from .composite_plan import CompositePlanViewModel
 
 
 class CommandInputViewModel:
@@ -145,37 +144,6 @@ class CommandParameterViewModel:
         )
 
 
-class PersonViewModel:
-    """View model for ``Person``."""
-
-    def __init__(self, name: str, email: str, affiliation: Optional[str]) -> None:
-        self.name = name
-        self.email = email
-        self.affiliation = affiliation
-
-    @classmethod
-    def from_person(cls, person: Person):
-        """Create view model from ``Person``.
-
-        Args:
-            person(Person): The person to convert.
-        Returns:
-            View model for person
-        """
-        return cls(name=person.name, email=person.email, affiliation=person.affiliation)
-
-    def __str__(self) -> str:
-        email = affiliation = ""
-
-        if self.email:
-            email = f" <{self.email}>"
-
-        if self.affiliation:
-            affiliation = f" [{self.affiliation}]"
-
-        return f"{self.name}{email}{affiliation}"
-
-
 class PlanViewModel:
     """A view model for a ``Plan``."""
 
@@ -232,7 +200,7 @@ class PlanViewModel:
         )
 
 
-def plan_view(workflow: AbstractPlan) -> Union["CompositePlanViewModel", PlanViewModel]:
+def plan_view(workflow: AbstractPlan) -> Union[CompositePlanViewModel, PlanViewModel]:
     """Convert an ``CompositePlan`` or ``Plan`` to a ``ViewModel``.
 
     Args:
@@ -241,7 +209,6 @@ def plan_view(workflow: AbstractPlan) -> Union["CompositePlanViewModel", PlanVie
     Returns:
         View model for converted Plan.
     """
-    from .composite_plan import CompositePlanViewModel
 
     if isinstance(workflow, CompositePlan):
         return CompositePlanViewModel.from_composite_plan(workflow)
